Extract still capture builder into camera.request
